+++ js/pirsch-events.js
@@ -1,12 +1,35 @@
-(function() {
+(function () {
     "use strict";
 
     // respect Do-Not-Track
-    if(navigator.doNotTrack === "1" || localStorage.getItem("disable_pirsch")) {
+    if (navigator.doNotTrack === "1" || localStorage.getItem("disable_pirsch")) {
         return;
     }
 
     const script = document.querySelector("#pirscheventsjs");
+
+    // include pages
+    try {
+        const include = script.getAttribute("data-include");
+        const paths = include ? include.split(",") : [];
+
+        if (paths.length) {
+            let found = false;
+
+            for (let i = 0; i < paths.length; i++) {
+                if (new RegExp(paths[i]).test(location.pathname)) {
+                    found = true;
+                    break;
+                }
+            }
+
+            if (!found) {
+                return;
+            }
+        }
+    } catch (e) {
+        console.error(e);
+    }
 
     // exclude pages
     try {
@@ -18,22 +41,22 @@
                 return;
             }
         }
-    } catch(e) {
+    } catch (e) {
         console.error(e);
     }
 
     // register event function
     const endpoint = script.getAttribute("data-endpoint") || "/pirsch/event";
-    window.pirsch = function(name, options) {
-        if(typeof name !== "string" || !name) {
+    window.pirsch = function (name, options) {
+        if (typeof name !== "string" || !name) {
             return Promise.reject("The event name for Pirsch is invalid (must be a non-empty string)! Usage: pirsch('event name', {duration: 42, meta: {key: 'value'}})");
         }
 
         return new Promise((resolve, reject) => {
             const meta = options && options.meta ? options.meta : {};
 
-            for(let key in meta) {
-                if(meta.hasOwnProperty(key)) {
+            for (let key in meta) {
+                if (meta.hasOwnProperty(key)) {
                     meta[key] = String(meta[key]);
                 }
             }
@@ -42,7 +65,7 @@
             req.open("POST", endpoint);
             req.setRequestHeader("Content-Type", "application/json;charset=UTF-8");
             req.onload = () => {
-                if(req.status >= 200 && req.status < 300) {
+                if (req.status >= 200 && req.status < 300) {
                     resolve(req.response);
                 } else {
                     reject(req.statusText);