+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Changelog
 
+## 1.1.0
+
+* updated SDK
+* added IP header configuration
+* added allowed subnets configuration
+* added whitelisting pages to scripts
+
 ## 1.0.0
 
 * updated SDK and other dependencies